+++ api/src/main/java/ru/gb/market/api/order/OrderItemDto.java
@@ -1,16 +1,18 @@
 package ru.gb.market.api.order;
 
+import java.math.BigDecimal;
+
 public class OrderItemDto {
     private Long productId;
     private String productTitle;
     private int count;
-    private int pricePerProduct;
-    private int price;
+    private BigDecimal pricePerProduct;
+    private BigDecimal price;
 
     public OrderItemDto() {
     }
 
-    public OrderItemDto(Long productId, String productTitle, int count, int pricePerProduct, int price) {
+    public OrderItemDto(Long productId, String productTitle, int count, BigDecimal pricePerProduct, BigDecimal price) {
         this.productId = productId;
         this.productTitle = productTitle;
         this.count = count;
@@ -34,19 +36,19 @@ public class OrderItemDto {
         this.count = count;
     }
 
-    public int getPricePerProduct() {
+    public BigDecimal getPricePerProduct() {
         return pricePerProduct;
     }
 
-    public void setPricePerProduct(int pricePerProduct) {
+    public void setPricePerProduct(BigDecimal pricePerProduct) {
         this.pricePerProduct = pricePerProduct;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
